inject will now inject interface implementors as dependencies.
If no concrete type is matched for the given type, and the type is an
interface, then inject will iterate the values map to search for any
implementors of that interface before moving to the parent injector.

This will be a performance hit for certain cases where the requested
dependencies exist concretely in the parent injector. The good news is
that this should solve issues with rewritten import paths and should
allow a looser injection scheme.