Use godocdown to create a github-friendly readme
This commit adds a script 'update_readme.sh' that generates
github-friendly markdown using godocdown. When the documentation for the
readme for github should be updated because the package or package's
documentation changes, run:

sh update_readme.sh

to update README.md.  This commit is meant to be the first step in
addressing #3.